+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/services/CustomItemDataService.java
@@ -3,6 +3,7 @@ package io.github.thebusybiscuit.slimefun4.core.services;
 import java.util.Optional;
 
 import org.bukkit.Keyed;
+import org.bukkit.Material;
 import org.bukkit.NamespacedKey;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.ItemMeta;
@@ -65,7 +66,7 @@ public class CustomItemDataService implements PersistentDataService, Keyed {
      * @return An {@link Optional} describing the result
      */
     public Optional<String> getItemData(ItemStack item) {
-        if (item == null || !item.hasItemMeta()) {
+        if (item == null || item.getType() == Material.AIR || !item.hasItemMeta()) {
             return Optional.empty();
         }