+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 * items which cannot be distributed by a Cargo Net will be dropped on the ground now instead of getting deleted
 * Small performance improvements to the Cargo Net
 * Slimefun no longer supports CraftBukkit
+* Item Energy is now also stored persistently via NBT
 
 #### Fixes
 * Fixed #2005
@@ -56,6 +57,7 @@
 * Fixed #1855
 * Fixed some issues with AsyncWorldEdit
 * Fixed some problems with unregistered or fake worlds
+* Fixed a rare concurrency issue with world saving
 
 ## Release Candidate 13 (16 Jun 2020)
 https://thebusybiscuit.github.io/builds/TheBusyBiscuit/Slimefun4/stable/#13